Fix bug #57077
Fix the problem with table style preview rendering. Now the class CDocumentContent can be completely independent of the parent class, so we can use CDocumentContent without specifying the parent. Also from now on there will no longer be a situation where we do not have a theme, color map, numbering, styles, if the global class is not specified we will use defaults.