Hotfix/v7.2.2 (#3076)
* Fix/pre7.2.2 (#3064)

* Fix bug 57481

* [bug] fix bug 59474:
fix tables conversion at copy/paste from the doc to the presentation editor

* Fix bug #59487

Fix the problem with rendering graphemes with several dicritical characters

* fix typo

* [de] Replace diacritical marks term with forced combined marks

Before that, we had special processing of diacritical marks. From now we will consider such marks as combined marks despite the fact that the text shaper doesn't recognize this marks as combined. And all processing will be the same as with normal combined marks

* [de] Remove old scheme with processing diacritical marks

* [de] Add simple tests for text shaper

* Fix bug #59461

Fix the problem with shaping text in case when characters with "inherited" script are present

* [de] Fix rendering non supported characters in the current font

Co-authored-by: Sergey Luzyanin <Sergey.Luzyanin@onlyoffice.com>
Co-authored-by: KirillovIlya <ilya.kirillov@onlyoffice.com>

* [bug] fix bug 59529

* [se] Fix bug 59549

Co-authored-by: papacarlo <builder@onlyoffice.com>
Co-authored-by: Sergey Luzyanin <Sergey.Luzyanin@onlyoffice.com>
Co-authored-by: KirillovIlya <ilya.kirillov@onlyoffice.com>
Co-authored-by: GoshaZotov <Igor.Zotov@onlyoffice.com>